+++ FootScout.WebAPI/Models/DTOs/SalaryRangeDTO.cs
@@ -4,7 +4,7 @@ namespace FootScout.WebAPI.Entities
 {
     public class SalaryRangeDTO
     {
-        public int Min { get; set; }
-        public int Max { get; set; }
+        public double Min { get; set; }
+        public double Max { get; set; }
     }
 }
